+++ src/Components/Pages/Visualizer/Ancilliary/Computation/TransInfo.js
@@ -63,3 +63,38 @@ export const computeTableData = (messageHistory) => {
 
     return { data, totalPctFaulty, totalHistLength, noPrimaryCnt };
 };
+
+export const truncData = (data, currentTransaction) => {
+    if (data === null || data === undefined) return {};
+    const filteredData = Object.fromEntries(
+        Object.entries(data).map(([key, value]) => {
+            if (value) {
+                const { replicaDetails, ...rest } = value;
+                return [key, rest];
+            }
+            return undefined;
+        }).filter(entry => entry !== undefined)
+    );
+
+    const entries = Object.entries(filteredData);
+
+    let selectedEntries = [];
+
+    if (filteredData[currentTransaction]) {
+        const currentTransactionEntry = entries.find(([key]) => key == currentTransaction);
+
+        selectedEntries.push(currentTransactionEntry);
+        const remainingEntries = entries.filter(([key]) => key != currentTransaction);
+        selectedEntries = selectedEntries.concat(remainingEntries.sort(() => Math.random() - 0.5).slice(0, 9));
+
+    } else {
+        selectedEntries = entries.sort(() => Math.random() - 0.5).slice(0, 10);
+    }
+    const result = {};
+    selectedEntries.forEach(([key, value]) => {
+        result[key] = value;
+    });
+    return result;
+};
+
+
