+++ src/Styles/App.css
@@ -1061,6 +1061,9 @@ video {
 .border-b {
   border-bottom-width: 1px;
 }
+.border-b-1p {
+  border-bottom-width: 1px;
+}
 .border-b-2 {
   border-bottom-width: 2px;
 }
@@ -1920,11 +1923,6 @@ code {
   border-color: rgb(255 255 255 / var(--tw-border-opacity));
 }
 
-:is([data-theme="dark"] .dark\:bg-black) {
-  --tw-bg-opacity: 1;
-  background-color: rgb(0 0 0 / var(--tw-bg-opacity));
-}
-
 :is([data-theme="dark"] .dark\:bg-blue-300) {
   --tw-bg-opacity: 1;
   background-color: rgb(10 10 10 / var(--tw-bg-opacity));
@@ -2073,11 +2071,6 @@ code {
   content: var(--tw-content);
   --tw-bg-opacity: 1;
   background-color: rgb(248 113 113 / var(--tw-bg-opacity));
-}
-
-:is([data-theme="dark"] .dark\:hover\:bg-black:hover) {
-  --tw-bg-opacity: 1;
-  background-color: rgb(0 0 0 / var(--tw-bg-opacity));
 }
 
 :is([data-theme="dark"] .dark\:hover\:bg-gray-600:hover) {
